Refactor: Consolidate Monte Carlo simulation engine
