+++ src/training/steps/backtesting/UNIFIED_CONFIG_EXAMPLES.md
@@ -337,7 +337,7 @@ results = await engine.execute_backtest(data, signals)
 
 ### **Monte Carlo Engine**
 ```python
-from src.training.steps.backtesting.real_monte_carlo_engine import RealMonteCarloEngine
+from src.utils.nas_tas.monte_carlo_engine import UnifiedMonteCarloEngine
 
 # Create configuration
 config = (create_config()
