+++ src/utils/ml_common/optimization/tas/backtesting/__init__.py
@@ -16,7 +16,7 @@ from .walk_forward_analysis import WalkForwardAnalyzer, WalkForwardConfig
 from .performance_attribution import PerformanceAttributor, AttributionConfig
 from .risk_analysis import RiskAnalyzer, RiskConfig
 from .scenario_testing import ScenarioTester, ScenarioConfig
-from .monte_carlo import MonteCarloSimulator, MonteCarloConfig
+from src.utils.nas_tas.monte_carlo_engine import UnifiedMonteCarloEngine as MonteCarloSimulator, MonteCarloConfig
 from .data_manager import BacktestingDataManager, DataConfig
 
 __all__ = [
